discord-bot/requirements.txt
discord.py>=2.3.2
aiohttp>=3.8.5
python-dotenv>=1.0.0
redis>=4.5.0
certifi>=2023.7.22
